Refactor Wasm impl to `hewn` engine and impl for snake
