+++ examples/snake/src/main.rs
@@ -1,19 +1,14 @@
 mod game;
 use hewn::game_object::Coordinate;
-
 #[cfg(not(target_arch = "wasm32"))]
 use hewn::{
     display::BaseDisplay, initialize_terminal_io, runtime::TerminalRuntime, TerminalRenderer,
 };
-fn main() {
-    play_snake_in_terminal();
-}
 
 const SCREEN_WIDTH: u16 = 30;
 const SCREEN_HEIGHT: u16 = 25;
 
-#[cfg(not(target_arch = "wasm32"))]
-fn play_snake_in_terminal() {
+fn main() {
     use crate::game::default;
 
     let (stdout, stdin) = initialize_terminal_io();
